06.01.2004 Tha Tab-Key switches now between Docs
git-svn-id: svn://scribus.net/trunk/Scribus@146 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/mpalette.h
@@ -208,7 +208,7 @@ public:
     ScribusDoc *doc;
     double Umrech;
     bool LMode;
-		int RoVal;
+		double RoVal;
 
 public slots:
 		void SetDoc(ScribusDoc *d);
